feat: retrospective window with basic react together
The retrospective window is already created and acessible with a sticky notes board
with 4 categories to put sticky notes in. It allows to like, dislike and comment on them.
It's also possible to chat with users as well as zoom in or zoom out in the board. Although
comments are possible they don't totally work as i am trying to explore better
possibilities than the one used.